# BackEnd/src/main/java/com/example/demo/models/SolarPanel.java
package com.example.demo.models;


import jakarta.persistence.*;
import lombok.Data;

@Data
@Entity
@Table(name = "solarPanel")
public class SolarPanel {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private String model;
    private Float nominal_power;
    private Float maximum_voltage;
    private Float maximum_current;
    private Float height;
    private Float width;
    private Float weight;
    private String type_cell;
    private Float price;
    private Integer quantity;
    @Lob
    @Column(name = "image", length = 500000000)
    private String image;
}