# BackEnd/src/main/java/com/example/demo/models/SystemFixing.java
package com.example.demo.models;


import jakarta.persistence.*;
import lombok.Data;

@Data
@Entity
@Table(name = "systemeFixing")
public class SystemFixing {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private String model;
    private String type;
    private String materiel;
    private Float charge;
    private Float height;
    private Float width;
    private String installation_method;
    private String adaptability;
    private Float price;
    private Integer quantity;
    @Lob
    @Column(name = "image", length = 500000000)
    private String image;
}