# FrontEnd/src/openApi/openApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/worker/update": {
      "put": {
        "tags": [
          "worker-controller"
        ],
        "operationId": "updateWorker",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Worker"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/systemFixing/update": {
      "put": {
        "tags": [
          "system-fixing-controller"
        ],
        "operationId": "updateSystemFixing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemFixing"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/supplier/update": {
      "put": {
        "tags": [
          "supplier-controller"
        ],
        "operationId": "updateSuplier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Supplier"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/solarPanel/update": {
      "put": {
        "tags": [
          "solar-panel-controller"
        ],
        "operationId": "updateSolarPanel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolarPanel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/meter/update": {
      "put": {
        "tags": [
          "meter-controller"
        ],
        "operationId": "updateMeter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/inverter/update": {
      "put": {
        "tags": [
          "inverter-controller"
        ],
        "operationId": "updateInverter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inverter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/worker/add": {
      "post": {
        "tags": [
          "worker-controller"
        ],
        "operationId": "addWorker",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Worker"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/systemFixing/add": {
      "post": {
        "tags": [
          "system-fixing-controller"
        ],
        "operationId": "ajouter_systemeFixation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemFixing"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/supplier/add": {
      "post": {
        "tags": [
          "supplier-controller"
        ],
        "operationId": "addSupplier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Supplier"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/solarPanel/add": {
      "post": {
        "tags": [
          "solar-panel-controller"
        ],
        "operationId": "addSolarPanel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolarPanel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/meter/delete/{id}": {
      "post": {
        "tags": [
          "meter-controller"
        ],
        "operationId": "deleteMeterById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/meter/add": {
      "post": {
        "tags": [
          "meter-controller"
        ],
        "operationId": "addMeter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/inverter/add": {
      "post": {
        "tags": [
          "inverter-controller"
        ],
        "operationId": "addInverter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inverter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/worker/get/one": {
      "get": {
        "tags": [
          "worker-controller"
        ],
        "operationId": "getWorkerByEmail",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Worker"
                }
              }
            }
          }
        }
      }
    },
...
    "/worker/delete/{id}": {
      "delete": {
        "tags": [
          "worker-controller"
        ],
        "operationId": "deleteWorker",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/systemFixing/delete/{id}": {
      "delete": {
        "tags": [
          "system-fixing-controller"
        ],
        "operationId": "deleteSystemFixing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/supplier/delete/{id}": {
      "delete": {
        "tags": [
          "supplier-controller"
        ],
        "operationId": "deleteSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/solarPanel/delete/{id}": {
      "delete": {
        "tags": [
          "solar-panel-controller"
        ],
        "operationId": "deleteSolarPanelById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/inverter/delete/{id}": {
      "delete": {
        "tags": [
          "inverter-controller"
        ],
        "operationId": "deleteInverterById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Activity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "start_date": {
            "type": "string",
            "format": "date-time"
          },
          "end_date": {
            "type": "string",
            "format": "date-time"
          },
          "text": {
            "type": "string"
          },
          "construction": {
            "type": "integer",
            "format": "int64"
          },
          "vehicule": {
            "type": "integer",
            "format": "int64"
          },
          "spots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Spot"
            }
          },
          "workers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Worker"
            }
          }
        }
      },
      "Person": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "Spot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Worker": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "generaleInfo": {
            "$ref": "#/components/schemas/Person"
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          }
        }
      },
      "SystemFixing": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "materiel": {
            "type": "string"
          },
          "charge": {
            "type": "number",
            "format": "float"
          },
          "height": {
            "type": "number",
            "format": "float"
          },
          "width": {
            "type": "number",
            "format": "float"
          },
          "installation_method": {
            "type": "string"
          },
          "adaptability": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string"
          }
        }
      },
      "Supplier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "SolarPanel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "nominal_power": {
            "type": "number",
            "format": "float"
          },
          "maximum_voltage": {
            "type": "number",
            "format": "float"
          },
          "maximum_current": {
            "type": "number",
            "format": "float"
          },
          "height": {
            "type": "number",
            "format": "float"
          },
          "width": {
            "type": "number",
            "format": "float"
          },
          "weight": {
            "type": "number",
            "format": "float"
          },
          "type_cell": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string"
          }
        }
      },
      "Meter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "connexion_type": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "format": "float"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string"
          }
        }
      },
      "Inverter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "nominal_power": {
            "type": "number",
            "format": "float"
          },
          "tension": {
            "type": "number",
            "format": "float"
          },
          "frequency": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "phase_number": {
            "type": "integer",
            "format": "int32"
          },
          "maximum_circuit_voltage": {
            "type": "number",
            "format": "float"
          },
          "minimal_circuit_current": {
            "type": "number",
            "format": "float"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string"
          }
        }
      },
...
    }
  }
}